:lipstick: Rename formatters -> writters
