+++ Makefile
@@ -12,7 +12,7 @@ SRCS= src/main.cpp \
 	src/uuid_consumer/kafka_uuid_consumer.cpp \
 	src/uuid_counter/uuid_counter.cpp \
 	src/uuid_counters_db/uuid_counters_db.cpp \
-	src/formatters/json_kafka_message.cpp \
+	src/writters/json_kafka_message.cpp \
 	src/uuid_counters_monitor/kafka_leadership_monitor.cpp \
 	src/monitor_producer/kafka_monitor_producer.cpp
 OBJS= $(SRCS:.cpp=.o)
